refactor resume
