+++ tests/prompts/routes.ts
@@ -10,12 +10,16 @@ export const TEST_PROMPTS = {
       parts: [{ type: 'text', text: 'Why is the sky blue?' }],
     },
     OUTPUT_STREAM: [
-      '0:"It\'s "',
-      '0:"just "',
-      '0:"blue "',
-      '0:"duh! "',
-      'e:{"finishReason":"stop","usage":{"promptTokens":3,"completionTokens":10},"isContinued":false}',
-      'd:{"finishReason":"stop","usage":{"promptTokens":3,"completionTokens":10}}',
+      'data: {"type":"start-step"}',
+      'data: {"type":"text-start","id":"STATIC_ID"}',
+      'data: {"type":"text-delta","id":"STATIC_ID","delta":"It\'s "}',
+      'data: {"type":"text-delta","id":"STATIC_ID","delta":"just "}',
+      'data: {"type":"text-delta","id":"STATIC_ID","delta":"blue "}',
+      'data: {"type":"text-delta","id":"STATIC_ID","delta":"duh! "}',
+      'data: {"type":"text-end","id":"STATIC_ID"}',
+      'data: {"type":"finish-step"}',
+      'data: {"type":"finish"}',
+      'data: [DONE]',
     ],
   },
   GRASS: {
@@ -26,14 +30,17 @@ export const TEST_PROMPTS = {
       content: 'Why is grass green?',
       parts: [{ type: 'text', text: 'Why is grass green?' }],
     },
-
     OUTPUT_STREAM: [
-      '0:"It\'s "',
-      '0:"just "',
-      '0:"green "',
-      '0:"duh! "',
-      'e:{"finishReason":"stop","usage":{"promptTokens":3,"completionTokens":10},"isContinued":false}',
-      'd:{"finishReason":"stop","usage":{"promptTokens":3,"completionTokens":10}}',
+      'data: {"type":"start-step"}',
+      'data: {"type":"text-start","id":"STATIC_ID"}',
+      'data: {"type":"text-delta","id":"STATIC_ID","delta":"It\'s "}',
+      'data: {"type":"text-delta","id":"STATIC_ID","delta":"just "}',
+      'data: {"type":"text-delta","id":"STATIC_ID","delta":"green "}',
+      'data: {"type":"text-delta","id":"STATIC_ID","delta":"duh! "}',
+      'data: {"type":"text-end","id":"STATIC_ID"}',
+      'data: {"type":"finish-step"}',
+      'data: {"type":"finish"}',
+      'data: [DONE]',
     ],
   },
 };
